Updated to support 1.20.4
Fixed minor issue with en_gb translation.
Added 1.20.4 support for Fabric and NeoForge to latest version.
