--- 1.20.4/common/src/main/java/io/github/samthegamer39/blocks/PoleBlock.java
@@ -1,103 +1,0 @@
-package io.github.samthegamer39.blocks;
-
-import net.minecraft.core.BlockPos;
-import net.minecraft.core.Direction;
-import net.minecraft.world.item.context.BlockPlaceContext;
-import net.minecraft.world.level.BlockGetter;
-import net.minecraft.world.level.LevelAccessor;
-import net.minecraft.world.level.block.Block;
-import net.minecraft.world.level.block.SimpleWaterloggedBlock;
-import net.minecraft.world.level.block.state.BlockState;
-import net.minecraft.world.level.block.state.StateDefinition;
-import net.minecraft.world.level.block.state.properties.BlockStateProperties;
-import net.minecraft.world.level.block.state.properties.BooleanProperty;
-import net.minecraft.world.level.material.FluidState;
-import net.minecraft.world.level.material.Fluids;
-import net.minecraft.world.phys.shapes.CollisionContext;
-import net.minecraft.world.phys.shapes.VoxelShape;
-
-/**
- * This class is designed for the mod's pole blocks.
- */
-public class PoleBlock extends Block implements SimpleWaterloggedBlock {
-
-    //Controls the shape of the collision/render block.
-    private static final VoxelShape SHAPE = Block.box(7, 0, 7, 9, 16, 9);
-    public static final BooleanProperty WATERLOGGED = BlockStateProperties.WATERLOGGED;
-
-    /**
-     * Constructor for blocks using the PoleBlock class. These blocks have two properties, HORIZONTAL_FACING and WATERLOGGED.
-     * @param properties Properties
-     */
-    public PoleBlock(Properties properties) {
-        super(properties);
-        this.registerDefaultState(this.defaultBlockState()
-                .setValue(WATERLOGGED, false));
-    }
-
-    /**
-     * Controls the state of the block when placed.
-     * @param context ItemPlacementContext
-     * @return Returns the state of the block as it should be when placed.
-     */
-    @Override
-    public BlockState getStateForPlacement(BlockPlaceContext context) {
-        return this.defaultBlockState()
-                .setValue(WATERLOGGED, context.getLevel().getFluidState(context.getClickedPos()).getType() == Fluids.WATER); //.getOpposite() can be added to flip the model.
-    }
-
-    /**
-     * Adds the block's properties.
-     * @param builder StateDefinition.Builder
-     */
-    @Override
-    protected void createBlockStateDefinition(StateDefinition.Builder<Block, BlockState> builder) {
-        super.createBlockStateDefinition(builder);
-        builder.add(WATERLOGGED);
-    }
-
-    /**
-     * Gets the current WATERLOGGED state, presumably.
-     * @param state Current block's state
-     * @return Returns the current WATERLOGGED state, presumably.
-     */
-    @SuppressWarnings("deprecation") //Why is getFluidState deprecated? Is there a better alternative?
-    @Override
-    public FluidState getFluidState(BlockState state) {
-        return state.getValue(WATERLOGGED) ? Fluids.WATER.getSource(false) : super.getFluidState(state);
-    }
-
-    /**
-     * Controls the outline of blocks using this class.
-     * @param state Current block's state
-     * @param worldIn BlockGetter
-     * @param pos Current position
-     * @param context CollisionContext
-     * @return Block outline shape
-     */
-    @SuppressWarnings("deprecation") //Why is getShape() deprecated?
-    @Override
-    public VoxelShape getShape(BlockState state, BlockGetter worldIn, BlockPos pos, CollisionContext context) {
-        return SHAPE;
-    }
-
-    /**
-     * I don't know what this does exactly. It's probably important.
-     * @param state Current block's state
-     * @param direction Direction
-     * @param neighborState Neighboring block's state
-     * @param world WorldAccess
-     * @param position Current position
-     * @param neighborPos Neighboring block's position
-     * @return
-     */
-    @SuppressWarnings("deprecation") //Why is updateShape deprecated? Is there a better alternative?
-    @Override
-    public BlockState updateShape(BlockState state, Direction direction, BlockState neighborState, LevelAccessor world, BlockPos position, BlockPos neighborPos){
-        if (state.getValue(WATERLOGGED)) {
-            world.scheduleTick(position,Fluids.WATER, Fluids.WATER.getTickDelay(world));
-        }
-
-        return super.updateShape(state, direction, neighborState, world, position, neighborPos);
-    }
-}
