+++ 1.20.4/common/src/main/resources/assets/railroadblocksaddon/blockstates/old_whistle_sign.json
@@ -1,0 +1,8 @@
+{
+  "variants": {
+    "facing=north": { "model": "railroadblocksaddon:block/old_whistle_sign", "y": 180 },
+    "facing=south": { "model": "railroadblocksaddon:block/old_whistle_sign"},
+    "facing=west":  { "model": "railroadblocksaddon:block/old_whistle_sign", "y": 90 },
+    "facing=east":  { "model": "railroadblocksaddon:block/old_whistle_sign", "y": 270 }
+  }
+}
