+++ 1.20.4/common/src/main/resources/assets/railroadblocksaddon/lang/en_gb.json
@@ -7,6 +7,7 @@
   "item.railroadblocksaddon.iron_pole": "Iron Post",
   "item.railroadblocksaddon.rxr_advance_sign": "Level Crossing Warning Sign",
   "item.railroadblocksaddon.whistle_sign": "Whistle Sign",
+  "item.railroadblocksaddon.old_whistle_sign": "Old Whistle Sign",
   "item.railroadblocksaddon.wooden_crossbuck": "Wooden Crossbuck",
   "item.railroadblocksaddon.wooden_pole": "Wooden Post",
 
@@ -14,6 +15,7 @@
   "block.railroadblocksaddon.iron_pole": "Iron Post",
   "block.railroadblocksaddon.rxr_advance_sign": "Level Crossing Warning Sign",
   "block.railroadblocksaddon.whistle_sign": "Whistle Sign",
+  "block.railroadblocksaddon.old_whistle_sign": "Old Whistle Sign",
   "block.railroadblocksaddon.wooden_crossbuck": "Wooden Crossbuck",
   "block.railroadblocksaddon.wooden_pole": "Wooden Post",
