+++ 1.20.4/common/src/main/resources/assets/railroadblocksaddon/models/block/old_whistle_sign.json
@@ -1,0 +1,46 @@
+{
+	"credit": "Made with Blockbench",
+	"texture_size": [32, 32],
+	"textures": {
+		"0": "railroadblocksaddon:block/old_whistle",
+		"particle": "railroadblocksaddon:block/old_whistle"
+	},
+	"elements": [
+		{
+			"from": [4, 0, 5],
+			"to": [12, 14, 11],
+			"faces": {
+				"north": {"uv": [4, 0.5, 8, 7.5], "texture": "#0"},
+				"east": {"uv": [8, 0.5, 11, 7.5], "texture": "#0"},
+				"south": {"uv": [0, 0.5, 4, 7.5], "texture": "#0"},
+				"west": {"uv": [11, 0.5, 14, 7.5], "texture": "#0"},
+				"up": {"uv": [4, 10.5, 0, 7.5], "texture": "#0"},
+				"down": {"uv": [8, 7.5, 4, 10.5], "texture": "#0"}
+			}
+		},
+		{
+			"from": [5, 14, 5],
+			"to": [11, 15, 11],
+			"faces": {
+				"north": {"uv": [0, 0, 3, 0.5], "texture": "#0"},
+				"east": {"uv": [9, 0, 12, 0.5], "texture": "#0"},
+				"south": {"uv": [3, 0, 6, 0.5], "texture": "#0"},
+				"west": {"uv": [6, 0, 9, 0.5], "texture": "#0"},
+				"up": {"uv": [3, 10.5, 0, 7.5], "texture": "#0"},
+				"down": {"uv": [6, 7.5, 3, 10.5], "texture": "#0"}
+			}
+		},
+		{
+			"from": [6, 15, 5],
+			"to": [10, 16, 11],
+			"faces": {
+				"north": {"uv": [0, 0, 2, 0.5], "texture": "#0"},
+				"east": {"uv": [9, 0, 12, 0.5], "texture": "#0"},
+				"south": {"uv": [3, 0, 5, 0.5], "texture": "#0"},
+				"west": {"uv": [6, 0, 9, 0.5], "texture": "#0"},
+				"up": {"uv": [2, 10.5, 0, 7.5], "texture": "#0"},
+				"down": {"uv": [5, 7.5, 3, 10.5], "texture": "#0"}
+			}
+		}
+	]
+}
