+++ 1.20.4/common/src/main/resources/assets/railroadblocksaddon/models/item/old_whistle_sign.json
@@ -1,0 +1,3 @@
+{
+  "parent": "railroadblocksaddon:block/old_whistle_sign"
+}
