+++ 1.20.4/common/src/main/resources/data/minecraft/tags/blocks/mineable/axe.json
@@ -5,6 +5,7 @@
     "railroadblocksaddon:iron_pole",
     "railroadblocksaddon:rxr_advance_sign",
     "railroadblocksaddon:whistle_sign",
+    "railroadblocksaddon:old_whistle_sign",
     "railroadblocksaddon:wooden_crossbuck",
     "railroadblocksaddon:wooden_pole"
   ]
