+++ 1.20.4/common/src/main/resources/data/railroadblocksaddon/recipes/flag_green.json
@@ -1,9 +1,8 @@
 {
   "type": "minecraft:crafting_shaped",
   "pattern": [
-    "   ",
-    " SG",
-    " S "
+    "SG",
+    "S "
   ],
   "key": {
     "S": {
@@ -14,7 +13,8 @@
     }
   },
   "result": {
-    "id": "railroadblocksaddon:flag_green",
-    "count": 1
-  }
+    "item": "railroadblocksaddon:flag_green",
+    "count": 3
+  },
+  "group": "flags"
 }
