+++ 1.20.4/common/src/main/resources/data/railroadblocksaddon/recipes/iron_crossbuck.json
@@ -11,7 +11,7 @@
     }
   },
   "result": {
-    "id": "railroadblocksaddon:iron_crossbuck",
+    "item": "railroadblocksaddon:iron_crossbuck",
     "count": 3
   }
 }
