+++ 1.20.4/common/src/main/resources/data/railroadblocksaddon/recipes/old_whistle_sign.json
@@ -1,0 +1,15 @@
+{
+  "type": "minecraft:crafting_shapeless",
+  "ingredients": [
+    {
+      "item": "minecraft:black_dye"
+    },
+    {
+      "item": "minecraft:light_gray_concrete_powder"
+    }
+  ],
+  "result": {
+    "item": "railroadblocksaddon:old_whistle_sign",
+    "count": 1
+  }
+}
