+++ 1.20.4/common/src/main/resources/data/railroadblocksaddon/recipes/rxr_advance_sign.json
@@ -14,7 +14,7 @@
     }
   },
   "result": {
-    "id": "railroadblocksaddon:rxr_advance_sign",
+    "item": "railroadblocksaddon:rxr_advance_sign",
     "count": 3
   }
 }
