+++ 1.20.4/common/src/main/resources/data/railroadblocksaddon/recipes/wooden_crossbuck.json
@@ -14,7 +14,7 @@
     }
   },
   "result": {
-    "id": "railroadblocksaddon:wooden_crossbuck",
+    "item": "railroadblocksaddon:wooden_crossbuck",
     "count": 3
   }
 }
