+++ 1.20.4/fabric/build.gradle
@@ -44,13 +44,17 @@ processResources {
                 fabric_version_range: fabric_loader_version_range,
                 fabric_api_version: fabric_api_version,
                 architectury_version_range: architectury_api_version_ranged,
+                java: java_version,
                 mod_id: mod_id,
                 mod_name: mod_name,
                 mod_license: license,
                 mod_version: mod_version,
                 mod_authors: mod_authors,
                 mod_credits: mod_credits,
-                mod_description: mod_description
+                mod_description: mod_description,
+                mod_homepage: mod_homepage,
+                mod_sources: mod_sources,
+                pack_format: pack_format
     }
 }
