+++ 1.20.4/gradle.properties
@@ -3,35 +3,37 @@ enabled_platforms = fabric,forge,quilt,neoforge
 license = MIT
 
 # Mod Properties
-maven_group = io.github.samthegamer39
-mod_authors = SamTheGamer39
+maven_group = dev.interurban
+mod_authors = Interurban Dev
 mod_id = railroadblocksaddon
 mod_name = Railroad Blocks
-mod_version = 0.5.0
+mod_version = 0.5.2
 mod_credits = Models made in Blockbench. Simplified Chinese translation by lonelyicer.
 mod_description = This mod adds railroad-related blocks to Minecraft!
-mod_homepage = https://github.com/SamTheGamer39/MinecraftRailroadBlocks
-mod_issues = https://github.com/SamTheGamer39/MinecraftRailroadBlocks/issues
-mod_sources = https://github.com/SamTheGamer39/MinecraftRailroadBlocks
-mod_update = https://samthegamer39.github.io/MinecraftRailroadBlocks/update.json
+mod_homepage = https://interurban.dev/MinecraftRailroadBlocks
+mod_issues = https://github.com/InterurbanDev/MinecraftRailroadBlocks/issues
+mod_sources = https://github.com/InterurbanDev/MinecraftRailroadBlocks
+mod_update = https://interurban.dev/MinecraftRailroadBlocks/update.json
 
 # Architectury
 architectury_version = 9.2.14
 architectury_api_version = 11.1.17
-architectury_api_version_range = [10.0.0, 12.0.0)
-architectury_api_version_ranged = >=10.0.0
+architectury_api_version_range = [9.0.0, 12.0.0)
+architectury_api_version_ranged = >=9.0.0
 
 # Common
+java_version = 17
 minecraft_version = 1.20.4
 minecraft_version_range = [1.20, 1.20.4]
 minecraft_version_ranged = 1.20.x
+pack_format = 22
 
 # Fabric
 # Check these on https://fabricmc.net/versions.html
 # ALWAYS check the fabric API mod page to see if the version is correct.
-fabric_api_version = 0.97.1+1.20.4
-fabric_loader_version = 0.15.11
-fabric_loader_version_range = >=0.15.11
+fabric_api_version = 0.97.2+1.20.4
+fabric_loader_version = 0.16.5
+fabric_loader_version_range = >=0.16.5
 
 # Forge
 forge_version = 49.1.0
@@ -43,7 +45,7 @@ neoforge_version_range = [20.2,20.5)
 neo_loader_version_range = [2,)
 
 # Quilt
-quilt_loader_version = 0.26.1-beta.1
+quilt_loader_version = 0.26.4
 quilted_fabric_api_version = 9.0.0-alpha.8+0.97.0-1.20.4
 
 # Gradle
