+++ 1.20.4/neoforge/src/main/resources/pack.mcmeta
@@ -1,6 +1,6 @@
 {
   "pack": {
     "description": "${mod_name}",
-    "pack_format": 34
+    "pack_format": 22
   }
 }
