+++ 1.20.4/quilt/build.gradle
@@ -36,7 +36,7 @@ dependencies {
     modImplementation "org.quiltmc:quilt-loader:$rootProject.quilt_loader_version"
 
     // Quilt Standard Libraries and QSL.
-    //modImplementation "org.quiltmc.quilted-fabric-api:quilted-fabric-api:$rootProject.quilted_fabric_api_version"
+    modImplementation "org.quiltmc.quilted-fabric-api:quilted-fabric-api:$rootProject.quilted_fabric_api_version"
 
     // Architectury API. This is optional, and you can comment it out if you don't need it.
     modImplementation("dev.architectury:architectury-fabric:$rootProject.architectury_api_version") {
