+++ 1.20.4/quilt/src/main/java/dev/interurban/quilt/RailroadBlocksQuilt.java
@@ -1,0 +1,21 @@
+package dev.interurban.quilt;
+
+import dev.interurban.RailroadBlocks;
+import org.quiltmc.loader.api.ModContainer;
+import org.quiltmc.qsl.base.api.entrypoint.ModInitializer;
+
+/**
+ * Main class for the RailroadBlocks mod. Initializes the mod and adds all associated blocks and items to Minecraft.
+ */
+public final class RailroadBlocksQuilt implements ModInitializer {
+
+	/**
+	 * This function initializes the mod and registers all blocks and items to the game.
+	 */
+	@Override
+	public void onInitialize(ModContainer mod) {
+		// Run common setup.
+		RailroadBlocks.init();
+		RailroadBlocks.LOGGER.info("Railroad Blocks successfully loaded.");
+	}
+}
