+++ 1.21/common/src/main/resources/assets/railroadblocksaddon/lang/en_gb.json
@@ -15,6 +15,7 @@
   "block.railroadblocksaddon.iron_pole": "Iron Post",
   "block.railroadblocksaddon.rxr_advance_sign": "Level Crossing Warning Sign",
   "block.railroadblocksaddon.whistle_sign": "Whistle Sign",
+  "block.railroadblocksaddon.old_whistle_sign": "Old Whistle Sign",
   "block.railroadblocksaddon.wooden_crossbuck": "Wooden Crossbuck",
   "block.railroadblocksaddon.wooden_pole": "Wooden Post",
