+++ 1.21/gradle.properties
@@ -7,7 +7,7 @@ maven_group = dev.interurban
 mod_authors = Interurban Dev
 mod_id = railroadblocksaddon
 mod_name = Railroad Blocks
-mod_version = 0.5.1
+mod_version = 0.5.2
 mod_credits = Models made in Blockbench. Simplified Chinese translation by lonelyicer.
 mod_description = This mod adds railroad-related blocks to Minecraft!
 mod_homepage = https://interurban.dev/MinecraftRailroadBlocks
